python_api/background-removal/requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
numpy
timm
kornia
scikit-image
huggingface_hub
transformers>=4.39.1
